requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.23.0
six==1.15.0
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
gunicorn==20.0.4
python-dotenv==0.14.0
